Update shell.c
So i added the file read in, however, there is an issue with the input variable not reading in anything after the first '\n' in the file.  We may need to attempt to store the information in a different variable type.  I am just not quite sure how we should do it.